# project-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Forge API
  description: |
    REST API for Forge, a file-system-backed project management tool for software teams.
    All request/response bodies use **snake_case** field names.
    Paginated list endpoints return `{ items, total, page, limit, total_pages }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/features:
    get:
      summary: List features (paginated)
      operationId: listFeatures
      tags: [Features]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - name: status
          in: query
          schema: { type: string }
          description: Comma-separated list of statuses (draft, deferred, ready, in-progress, review, done)
        - name: priority
          in: query
          schema: { type: string }
          description: Comma-separated list of priorities (critical, high, medium, low)
        - name: assignee
          in: query
          schema: { type: string }
          description: Filter by assignee name, or `__unassigned__` for unassigned features
        - name: points_min
          in: query
          schema: { type: number }
          description: Minimum points filter
        - name: points_max
          in: query
          schema: { type: number }
          description: Maximum points filter
        - name: tags
          in: query
          schema: { type: string }
          description: Comma-separated list of tags to filter by
        - name: sort
          in: query
          schema:
            type: string
            enum: [order, title, status, priority, modified]
            default: order
        - name: dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Paginated list of features
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFeatures"

    post:
      summary: Create a feature
      operationId: createFeature
      tags: [Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureInput"
      responses:
        "201":
          description: Feature created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/features/reorder:
    patch:
      summary: Reorder features
      operationId: reorderFeatures
      tags: [Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/features/{slug}:
    parameters:
      - $ref: "#/components/parameters/slug"

    get:
      summary: Get a feature by slug
      operationId: getFeature
      tags: [Features]
      responses:
        "200":
          description: Feature detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update a feature
      operationId: updateFeature
      tags: [Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureInput"
      responses:
        "200":
          description: Updated feature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete a feature
      operationId: deleteFeature
      tags: [Features]
      responses:
        "200":
          description: Feature deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  slug: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/features/{slug}/artifacts:
    parameters:
      - $ref: "#/components/parameters/slug"
    get:
      summary: List artifacts for a feature
      operationId: listFeatureArtifacts
      tags: [Features]
      responses:
        "200":
          description: List of artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/features/{slug}/artifacts/upload:
    parameters:
      - $ref: "#/components/parameters/slug"
    post:
      summary: Upload a file artifact
      operationId: uploadFeatureArtifact
      tags: [Features]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Artifact uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/features/{slug}/artifacts/create:
    parameters:
      - $ref: "#/components/parameters/slug"
    post:
      summary: Create an empty text artifact
      operationId: createFeatureArtifact
      tags: [Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtifactInput"
      responses:
        "201":
          description: Artifact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/features/{slug}/artifacts/{filename}:
    parameters:
      - $ref: "#/components/parameters/slug"
      - $ref: "#/components/parameters/filename"

    get:
      summary: Get artifact content
      description: |
        Returns JSON `{ name, content, type, mime_type, size }` for text files.
        Returns raw binary with appropriate Content-Type for binary files or when `?raw` is passed.
      operationId: getFeatureArtifactContent
      tags: [Features]
      parameters:
        - name: raw
          in: query
          schema: { type: boolean }
          description: Force raw binary response
      responses:
        "200":
          description: Artifact content (JSON for text, binary for non-text)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactContent"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Save artifact content
      operationId: saveFeatureArtifactContent
      tags: [Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveArtifactContentInput"
      responses:
        "200":
          description: Artifact saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete an artifact
      operationId: deleteFeatureArtifact
      tags: [Features]
      responses:
        "200":
          description: Artifact deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  slug: { type: string }
                  filename: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Issues ────────────────────────────────────────────────────────────────
...
components:
...
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
      description: Page number
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 25, minimum: 1, maximum: 100 }
      description: Items per page
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
    filename:
      name: filename
      in: path
      required: true
      schema: { type: string }

  # ── Responses ─────────────────────────────────────────────────────────────
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: true }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    ValidationError:
      description: Validation error (422)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, description: "Semicolon-separated list of field-level errors" }
    Conflict:
      description: Resource already exists (409)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }

  # ── Schemas ───────────────────────────────────────────────────────────────
  schemas:
...
    FeatureStatus:
      type: string
      enum: [draft, deferred, ready, in-progress, review, done]
...
    Priority:
      type: string
      enum: [critical, high, medium, low]

    # ── Feature ────────────────────────────────────────────────────────────
    Feature:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/FeatureStatus" }
        priority: { $ref: "#/components/schemas/Priority" }
        assignees:
          type: array
          items: { type: string }
        points: { type: number, nullable: true }
        sprint: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        created: { type: string, description: ISO 8601 date }
        modified: { type: string, description: ISO 8601 date }
        roadmap_card: { type: string, nullable: true, description: Linked roadmap card ID }
        body: { type: string }
        order: { type: number }
    PaginatedFeatures:
      type: object
      properties:
        features:
          type: array
          items: { $ref: "#/components/schemas/Feature" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }
    CreateFeatureInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        roadmap_card_id: { type: string, description: Roadmap card to link }
    UpdateFeatureInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        status: { $ref: "#/components/schemas/FeatureStatus" }
        priority: { $ref: "#/components/schemas/Priority" }
        assignees:
          type: array
          items: { type: string }
        points: { type: number, nullable: true }
        tags:
          type: array
          items: { type: string }
        body: { type: string }

    # ── Issue ──────────────────────────────────────────────────────────────
...
    ReorderInput:
      type: object
      required: [slugs]
      properties:
        slugs:
          type: array
          items: { type: string }
        offset: { type: number, description: Starting offset for the order values }

    # ── Artifacts ──────────────────────────────────────────────────────────
    Artifact:
      type: object
      properties:
        name: { type: string }
        size: { type: number }
        type: { type: string }
    ArtifactContent:
      type: object
      properties:
        name: { type: string }
        content: { type: string }
        type: { type: string }
        mime_type: { type: string }
        size: { type: number }
    CreateArtifactInput:
      type: object
      required: [filename]
      properties:
        filename: { type: string, minLength: 1 }
    SaveArtifactContentInput:
      type: object
      required: [content]
      properties:
        content: { type: string }

    # ── Search ─────────────────────────────────────────────────────────────
...